# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vraja Content Hub public contracts and critical editorial mutations",
    "version": "1.1.0",
    "description": "Local implementation. Additive v1 fields. No deployment or Android integration claim. Complete admin workflow in API-CONTRACT.md."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local. Verify the approved staging/production hostname before use."
    }
  ],
  "paths": {
...
    "/api/v1/books": {
      "get": {
        "summary": "Legacy books array or individual detail.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "description": "With id, returns one Book with pages.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful validated response. Only complete scope reconciliation can withdraw cached server-owned content.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/BookList"
                    },
                    {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Book"
                        },
                        {
                          "type": "object",
                          "properties": {
                            "pages": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/BookPage"
                              }
                            }
                          },
                          "required": [
                            "pages"
                          ],
                          "additionalProperties": true
                        }
                      ]
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request. Retain cached content and personal state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No eligible item or detail unavailable. Retain cached content and personal state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version or snapshot changed. Refresh and retry. Retain cached content and personal state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Request failed. Retain cached content and personal state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency or schema unavailable. Retain cached content and personal state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/daily-preview": {
      "get": {
        "summary": "Admin diagnostics; detailed shapes in API-CONTRACT.md.",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "locale",
            "in": "query",
            "required": false,
            "description": "Requested, base-language, then English fallback.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "coverage",
            "in": "query",
            "required": false,
            "description": "",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful validated response. Only complete scope reconciliation can withdraw cached server-owned content.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request. Retain cached content and personal state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication or role denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authentication or role denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No eligible item or detail unavailable. Retain cached content and personal state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version or snapshot changed. Refresh and retry. Retain cached content and personal state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Request failed. Retain cached content and personal state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency or schema unavailable. Retain cached content and personal state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/scripture-verses": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Trusted super_admin mutation, transactional and audited.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScriptureImport"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated record or import preview/result.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Rejected; no mutation committed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Rejected; no mutation committed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Rejected; no mutation committed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Rejected; no mutation committed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Rejected; no mutation committed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Rejected; no mutation committed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Rejected; no mutation committed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "details": {},
          "retryable": {
            "type": "boolean"
          }
        },
        "required": [
          "message"
        ],
        "additionalProperties": true
      },
...
      "BookPage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "page_number": {
            "type": "integer",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "body_text": {
            "type": [
              "string",
              "null"
            ]
          },
          "image_url": {
            "type": "string"
          },
          "thumbnail_url": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "page_number",
          "title",
          "body_text",
          "image_url",
          "thumbnail_url"
        ],
        "additionalProperties": true
      },
      "Book": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "cover_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "sort_order": {
            "type": "integer"
          },
          "page_count": {
            "type": "integer",
            "minimum": 0
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "pages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookPage"
            }
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "cover_url",
          "sort_order",
          "created_at",
          "updated_at"
        ],
        "additionalProperties": true
      },
...
      "BookList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Book"
        }
      },
...
      "ScriptureImport": {
        "type": "object",
        "properties": {
          "scripture_id": {
            "type": "string"
          },
          "mode": {
            "enum": [
              "replace",
              "append",
              "merge"
            ]
          },
          "expected_version": {
            "type": "integer",
            "minimum": 0
          },
          "verses": {
            "type": "array",
            "maxItems": 20000,
            "items": {
              "type": "object",
              "properties": {
                "canto": {
                  "type": [
                    "integer",
                    "null"
                  ]
                },
                "chapter": {
                  "type": "integer",
                  "minimum": 1
                },
                "verse": {
                  "type": "integer",
                  "minimum": 1
                },
                "translation": {
                  "type": "string"
                }
              },
              "required": [
                "chapter",
                "verse",
                "translation"
              ],
              "additionalProperties": true
            }
          },
          "preview": {
            "type": "boolean"
          },
          "reviewed_correction": {
            "type": "boolean"
          },
          "removed_resolution": {
            "enum": [
              "retain",
              "delete_unreferenced"
            ]
          },
          "confirm_removed_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "scripture_id",
          "verses"
        ],
        "additionalProperties": true
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}